+++ RevenueCat/Scripts/Error.cs
@@ -1,0 +1,15 @@
+using System;
+using System.Diagnostics.CodeAnalysis;
+
+public partial class Purchases
+{
+    [Serializable]
+    [SuppressMessage("ReSharper", "NotAccessedField.Global")]
+    public class Error
+    {
+        public string message;
+        public int code;
+        public string underlyingErrorMessage;
+        public string readableErrorCode;
+    }
+}
